requirements.txt
requests>=2.25.0
openai>=0.27.0
python-dotenv>=0.19.0
